+++ bpm-engine-api/src/main/java/io/takari/bpm/model/diagram/Shape.java
@@ -3,13 +3,13 @@ package io.takari.bpm.model.diagram;
 import java.io.Serializable;
 
 public class Shape implements Serializable {
-	
-	private static final long serialVersionUID = 1L;
+
+    private static final long serialVersionUID = 1L;
 
     private final String id;
     private final String elementId;
     private final Bounds bounds;
-    
+
     public Shape(String id, String elementId, Bounds bounds) {
         this.id = id;
         this.elementId = elementId;